bluetooth: controller: removed HCI ISO data packet length guard
Since the ISO data will now be fragmented by the ISO-AL, a guard on the
maximum length that the LL can handled should not be necessary.
LL_LENGTH_OCTETS_TX_MAX is defined based on CONFIG_BT_BUF_ACL_TX_SIZE,
which should not be the basis for ISO data.

Signed-off-by: Nirosharn Amarasinghe <niag@demant.com>