bluetooth: controller: Prevent HCI reset race with RX thread
When an HCI reset is issued, the HCI RX thread may be in the middle of
processing RX nodes. Normally this is not a problems as the RX/TX
cooperative threads are exclusive. However, if the RX thread yields via
a kernel functions like waiting for a semaphore or netbuffer, the RX
thread may have unfinished business when the reset is executed.
This means that when reset is done, the processing continues, and may
operate on invalidated data.

To prevent this, a new "processing" semaphore is introduced in the RX
thread. Before starting node processing, the semaphore is taken, and
given when processing is done.
The HCI reset function also takes this semahore, thereby preventing
reset starting in the middle of node processing. Whenever the RX thread
takes the semaphore, it checks for reset signal to see if a reset
happened while waiting. In this case the node is discarded.

Signed-off-by: Morten Priess <mtpr@oticon.com>