bluetooth: controller: Fix ISO-AL unframed burst discontinuity
For framed mode, BN>1, prevent ISO-AL from assigning non-consecutive
payload numbers. Numbering within a burst must be consecutive to allow
correct transmission and flush point handling in the link layer.

Previously we had a situation where if a burst of SDUs were released
across the ISO interval mark, it would have introduced a skew due by
moving SDUs to the next event and the burst would not appear consecutive
at the receiver. In this change, a burst of SDUs should remain
consecutive, however there is now a risk of losing data if there are
multiple SDUs per ISO interval and SDUs are not released consecutively
because the payload number only aligns with the event after BN number
of payloads. This solution, however, has more qualification tests
passing, and a new general solution to this problem should be found.

This change prevents "late" delivered SDUs from targeting PDUs in a
later event (skip payload numbers), but will request all payloads in a
burst transmitted in the same ISO event.

Adjust SDU target_event to take ISO event state into consideration.
Payloads will be requested postponed by an extra ISO event if outside
active ISO event "zone", as it is not known if a PDU can make the
deadline and be be delivered in event_count + 1.

Signed-off-by: Morten Priess <mtpr@oticon.com>