Bluetooth: controller: fix procedure collision handling
If an instant based remote procedure 'overtakes' a local ditto
the local procedure will be 'stopped' but collision flag would not
get set and ensure that a new local conflicting procedure
cannot be started before the remote is completed.
This can thus lead to 'collision'.

Fix by ensuring collision flag is set also in the above mentioned
scenario.

Signed-off-by: Erik Brockhoff <erbr@oticon.com>